Sprint 1 — React foundation
