+++ client/src/features/settings/page.tsx
@@ -1,0 +1,12 @@
+import { Card } from "../../shared/ui/Card";
+
+export function SettingsPage() {
+  return (
+    <div className="space-y-4">
+      <h1 className="text-xl font-semibold">Settings</h1>
+      <Card>
+        <p className="text-sm text-slate-400">Sprint 1 shell.</p>
+      </Card>
+    </div>
+  );
+}
