+++ client/src/index.css
@@ -2,9 +2,13 @@
 @tailwind components;
 @tailwind utilities;
 
-/* Optional global body styles */
-body {
-  @apply bg-slate-950 text-slate-100 antialiased;
+@layer base {
+  html {
+    @apply bg-slate-950 text-slate-100;
+  }
+  body {
+    @apply antialiased;
+  }
 }
 
 @keyframes savedFade {
@@ -102,14 +106,6 @@ body {
   }
 }
 
-.controls-enter {
-  animation: controlsFadeIn 120ms ease-out both;
-}
-
-.controls-exit {
-  animation: controlsFadeOut 120ms ease-in both;
-}
-
 .controls-wrap {
   position: relative;
   min-width: 170px; /* prevents layout jump; adjust if needed */
@@ -117,22 +113,51 @@ body {
   justify-content: flex-end;
 }
 
-.controls-layer {
+/* Base layer: hidden and non-interactive by default */
+.controls-layer,
+.content-layer {
   position: absolute;
   right: 0;
   top: 0;
   display: flex;
   gap: 0.5rem;
+
+  opacity: 0;
+  transform: translateY(2px);
+  pointer-events: none;
+
+  transition:
+    opacity 120ms ease,
+    transform 120ms ease;
+  will-change: opacity, transform;
 }
 
-.controls-layer--static {
+/* Enter = visible */
+.controls-enter {
+  opacity: 1;
+  transform: translateY(0);
+  pointer-events: auto;
+  z-index: 2;
+}
+
+/* Exit = hidden */
+.controls-exit {
+  opacity: 0;
+  transform: translateY(-2px);
+  pointer-events: none;
+  z-index: 1;
+}
+
+/* Static reserve layers */
+.controls-layer--static,
+.content-layer--static {
   position: relative;
+  pointer-events: none;
 }
 
 .content-wrap {
   position: relative;
   flex: 1;
-  min-height: 78px; /* prevents layout jump; adjust if needed */
 }
 
 .content-layer {
@@ -146,9 +171,24 @@ body {
   position: relative;
 }
 
-.delay-0 {
-  animation-delay: 0ms;
+@keyframes editGlow {
+  0% {
+    box-shadow: 0 0 0 0 rgba(20, 184, 166, 0);
+  }
+  40% {
+    box-shadow: 0 0 0 3px rgba(20, 184, 166, 0.25);
+  }
+  100% {
+    box-shadow: 0 0 0 0 rgba(20, 184, 166, 0);
+  }
 }
-.delay-30 {
-  animation-delay: 30ms;
+
+.edit-glow {
+  animation: editGlow 260ms ease-out both;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .edit-glow {
+    animation: none;
+  }
 }
