+++ client/src/shared/ui/Input.tsx
@@ -9,8 +9,17 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
       <input
         ref={ref}
         {...props}
-        className={`w-full rounded border bg-white text-black placeholder:text-slate-400 px-3 py-2 focus:outline-none focus:ring-2 focus:ring-teal-500 ${className}`}
+        className={[
+          "w-full rounded border border-slate-300 bg-white text-black px-3 py-2",
+          // Left-only focus indicator
+          "focus:outline-none focus:ring-0",
+          "focus:border-teal-500",
+          "focus:shadow-[inset_4px_0_0_0_rgba(20,184,166,1)]",
+          className,
+        ].join(" ")}
       />
     );
   },
 );
+
+Input.displayName = "Input";
