+++ client/src/shared/ui/Select.tsx
@@ -14,7 +14,14 @@ export function Select<T extends string>({
   return (
     <select
       {...props}
-      className={`w-full rounded border border-slate-300 bg-white text-black px-3 py-2 focus:outline-none focus:ring-2 focus:ring-teal-500 ${className}`}
+      className={[
+        "w-full rounded border border-slate-300 bg-white text-black px-3 py-2",
+        // Left-only focus indicator
+        "focus:outline-none focus:ring-0",
+        "focus:border-teal-500",
+        "focus:shadow-[inset_4px_0_0_0_rgba(20,184,166,1)]",
+        className,
+      ].join(" ")}
     >
       {children}
     </select>
