+++ client/src/shared/ui/Spinner.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 type SpinnerProps = {
   size?: "sm" | "md";
   className?: string;
